# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Manage a WireGuard VPN concentrator (interfaces \u0026 peers) without touching the CLI.\nKeys, client configs and QR codes are generated server-side; peers are pushed to the kernel via netlink.",
        "title": "VPN Manager API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/interfaces/{id}/peers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "peers"
                ],
                "summary": "List peers of an interface",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Interface ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Server auto-generates keys and IP when omitted",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "peers"
                ],
                "summary": "Add a peer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Interface ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Peer payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePeerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/peers/{peerId}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "peers"
                ],
                "summary": "Update a peer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Peer ID",
                        "name": "peerId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Peer payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePeerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "peers"
                ],
                "summary": "Delete a peer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Peer ID",
                        "name": "peerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/peers/{peerId}/config": {
            "get": {
                "description": "Returns the wg-quick .conf for this peer as plain text",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "peers"
                ],
                "summary": "Get client config",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Peer ID",
                        "name": "peerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/peers/{peerId}/qrcode": {
            "get": {
                "description": "Returns a PNG QR code of the peer's wg-quick config",
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "peers"
                ],
                "summary": "Get client config QR code",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Peer ID",
                        "name": "peerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {},
                "success": {
                    "type": "boolean"
                },
                "warning": {
                    "description": "Warning is set when the primary operation succeeded and was persisted, but a\nsecondary best-effort step degraded (e.g. the WireGuard kernel reconcile or\na VPN runtime apply failed). The status stays 2xx because the resource change\nwas committed; the warning carries the degradation detail for the client.",
                    "type": "string"
                }
            }
        },
...
        "dto.CreatePeerRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "assigned_ip": {
                    "type": "string",
                    "example": "10.8.0.2"
                },
                "client_allowed_ips": {
                    "type": "string",
                    "example": "0.0.0.0/0, ::/0"
                },
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1,
                    "example": "laptop-andi"
                },
                "persistent_keepalive": {
                    "type": "integer",
                    "maximum": 65535,
                    "minimum": 0,
                    "example": 25
                },
                "public_key": {
                    "type": "string",
                    "example": ""
                },
                "use_preshared_key": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.UpdatePeerRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "client_allowed_ips": {
                    "type": "string",
                    "example": "0.0.0.0/0, ::/0"
                },
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1,
                    "example": "laptop-andi"
                },
                "persistent_keepalive": {
                    "type": "integer",
                    "maximum": 65535,
                    "minimum": 0,
                    "example": 25
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the token from /auth/login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}